[Feat] 홈파밍 결과 이미지 등록 컴포넌트 분리 (#143)
