+++ lib/main.dart
@@ -1,7 +1,6 @@
 import 'package:farmus/common/theme/farmus_theme_color.dart';
 import 'package:farmus/res/app_url/app_url.dart';
-import 'package:farmus/view/on_boarding/on_boarding_screen.dart';
-import 'package:farmus/view/vege_diary/vege_diary_screen.dart';
+import 'package:farmus/view/vege_delete/vege_delete_screen.dart';
 import 'package:farmus/view/vege_diary_write/vege_diary_write_screen.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/cupertino.dart';
@@ -35,7 +34,7 @@ Future<void> main() async {
   runApp(ProviderScope(
     child: MaterialApp(
       title: "팜어스",
-      home: const VegeDiaryWriteScreen(),
+      home: const VegeDeleteScreen(),
       debugShowCheckedModeBanner: false,
       localizationsDelegates: const [
         GlobalMaterialLocalizations.delegate,
