+++ lib/view/vege_diary_write/vege_diary_write_screen.dart
@@ -43,7 +43,9 @@ class VegeDiaryWriteScreen extends ConsumerWidget {
         ],
       ),
       body: const Column(
-        children: [],
+        children: [
+
+        ],
       ),
     );
   }
